feat: Remove deprecated integration tests and introduce new unit test directories
