+++ tests/unit/context_management/test_contextual_awareness_basic_integration.py
