+++ tests/unit/optimization/test_advanced_token_optimization_integration.py
