+++ tests/unit/optimization/test_proactive_optimization_integration.py
